removed stuff from the `v2` namespace, put responder in `api` and database manipulation in `db`
