+++ src/db/ai_history.rs
@@ -2,28 +2,15 @@ use crate::db::schema::ai_help_history;
 use crate::db::Pool;
 use crate::diesel::QueryDsl;
 use crate::{api::error::ApiError, settings::SETTINGS};
-use actix_rt::ArbiterHandle;
-use actix_web::{web::Data, HttpResponse};
+use actix_web::web::Data;
 use chrono::Utc;
 use diesel::{ExpressionMethods, RunQueryDsl};
 use std::ops::Sub;
 use std::time::Duration;
 
-pub async fn delete_old_ai_history(
-    pool: Data<Pool>,
-    arbiter: Data<ArbiterHandle>,
-) -> Result<HttpResponse, ApiError> {
-    if !arbiter.spawn(async move {
-        if let Err(e) = do_delete_old_ai_history(pool).await {
-            error!("{}", e);
-        }
-    }) {
-        return Ok(HttpResponse::InternalServerError().finish());
-    }
-    Ok(HttpResponse::Accepted().finish())
-}
-
-async fn do_delete_old_ai_history(pool: Data<Pool>) -> Result<(), ApiError> {
+/// This removes old AI history records from the database. It is meant to be called from a
+/// cron job calling the respective endpoint in the admin API.
+pub async fn do_delete_old_ai_history(pool: Data<Pool>) -> Result<(), ApiError> {
     let mut conn = pool.get()?;
     let history_deletion_period_in_sec = SETTINGS
         .ai
@@ -41,6 +28,8 @@ async fn do_delete_old_ai_history(pool: Data<Pool>) -> Result<(), ApiError> {
         ai_help_history::table.filter(ai_help_history::updated_at.lt(oldest_timestamp)),
     )
     .execute(&mut conn)?;
-    info!("Deleted old AI history: {affected_rows} old record(s) deleted.");
+    info!(
+        "Deleted old AI history before {oldest_timestamp}: {affected_rows} old record(s) deleted."
+    );
     Ok(())
 }
