+++ src/db/mod.rs
@@ -1,5 +1,6 @@
 pub mod ai_explain;
 pub mod ai_help;
+pub mod ai_history;
 pub mod documents;
 pub mod error;
 pub mod fxa_webhook;
