+++ src/db/v2/mod.rs
@@ -1,4 +1,3 @@
-pub mod ai_history;
 pub mod bcd_updates;
 pub mod collection_items;
 pub mod db_macros;
